Strengthening tests, fixing some bugs
Also making output a bit more structured, restricting decode where it
actually is generated, not all over the place, simplifying some methods
by accepting only specific types.

The current error handling is to default to empty/null/zero values on
error, so that one can check for valid output, but this won't fly as the
program gets more complex. We'll need a bit of exceptions in here soon.