Adding "regression" tests for aggregate.py
The test script basically runs it and checks for some consistency in the
output (correct categories, number of results and that it parses both
perf and Lulesh correctly).

This is a horrible way of testing but will do for now. We should create
a separate test directory and research some proper testing
infrastructure for python.