Improving statistical analysis description / plan
Updating the comment on aggregate.py to extend a bit more what I think
could be done by the script. Obviously, all of the actual analysis needs
to be done by modules, etc.

But the script should be able to understand the data (with or without
user help, via flags) and use the correct analysis across the correct
categories in the hierarchy.

I envisage two output modes:
 1. Dump all results, colouring / marking statistically relevant ones
 2. Only print a list of statistically relevant ones

(1) works for spreadsheets and CI while (2) works for manual runs and
profiling.