Simple clustering of data
Clustering is aimed at results that are known to be K-modal. The current
model works well for data without large outliers, but behaves
erradically when a distant outlier pull a centre all to itself and
defies the purpose.

This may call for an iterative approach, first finding global outliers,
then removing them and clustering, then finding cluster outliers, etc.

Also not implemented is any attempr to figure out the best number of
clusters rather than force users to hard code. This can be done in a
number of ways, including the pedestrian:
  for i in 1..k:
    best = min(best, kmeans(i).distances())
all the way to using SciKit:
  http://scikit-learn.org/stable/modules/clustering.html