Adding simple polynomial fit (least sq)
This is simply to look at progressions (increasing number of cores) and
see if the scalability goes up as the cores go. It also compares to an
optimal scalability and returns the average distance of the prediction
(could be geomean?).

Regardless, this will give us an idea of how to compare two scalability
results between different runs.