Improving fit's quality check a bit
Moving to a Chi=squared based, but not entirely. I don't want to add the
scipy dependency yet, as I just need some "ok" values to work with for
the parsing and organising. Once that part is done, all of these
analysis will be basically wrapped up scipy/numpy calls.