Adding support for data_string
This feature allows users to set what kind of analyses they want for
each category level and also what is the separator used in the log
files, for example '-' or '_', etc.

It's still not running the analysis on the categories yet. We may need
to also define what kind of comparison we'll do (across parent
categories or along same category.

Since the -d option is not mandatory, this doesn't change the behaviour
of the script right now.