Fixing analysis plugin structure
Now all analysis plugins derive from the same base class, AnalysisBase,
which guarantee the same behaviour for all passes. This makes it harder
to guarantee the needed parameters are set, as they're all options in a
dictionary, so I had to add a lot of exceptions for type, value, runtime
behaviour to be guaranteed.

It could do with a trim and better guarantees, but I think overall, this
is a recipe that can last longer on its own.